requirements.txt
numpy
tqdm
torch==1.10.1+cu113
torchvision==0.11.2+cu113
Pillow==9.0.1
